@@ -1,13 +1,28 @@
 module github.com/benitogf/auth
 
-go 1.15
+go 1.19
 
 require (
-	github.com/benitogf/katamari v0.0.0-20211230063327-83e07ad3fb0f
+	github.com/benitogf/katamari v0.0.0-20220924134414-1f73e9bf0582
 	github.com/benitogf/pivot v0.0.0-20210825050412-ec2fd40ab94f
+	github.com/goccy/go-json v0.9.11
 	github.com/golang-jwt/jwt v3.2.2+incompatible
 	github.com/gorilla/mux v1.8.0
-	github.com/pkg/expect v0.0.0-20191209053905-1fe4c9394a8a // indirect
-	github.com/stretchr/testify v1.7.0
+	github.com/stretchr/testify v1.8.0
 	golang.org/x/crypto v0.0.0-20211215153901-e495a2d5b3d3
+)
+
+require (
+	github.com/bclicn/color v0.0.0-20180711051946-108f2023dc84 // indirect
+	github.com/benitogf/coat v0.0.0-20200402073050-ff807656cbec // indirect
+	github.com/benitogf/jsonpatch v0.0.0-20210608090026-cebd07eba361 // indirect
+	github.com/cristalhq/base64 v0.1.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/felixge/httpsnoop v1.0.3 // indirect
+	github.com/gorilla/handlers v1.5.1 // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/pkg/expect v0.0.0-20191209053905-1fe4c9394a8a // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rs/cors v1.8.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
